requirements.txt
# LangChain Core
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# LangChain Integrations
langchain-openai==0.1.8

# LangGraph
langgraph==0.0.55

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.7.1

# Token Counting
tiktoken==0.6.0